# Optimize bicubic shader